Move util functions to folders where they are used
