+++ src/view/popup/__tests__/formatBytesString.test.ts
@@ -1,4 +1,4 @@
-import { formatBytesString } from "../formatBytesString";
+import { formatBytesString } from "../utils/formatBytesString";
 
 describe("formatBytesString", () => {
     it("format as bytes", () => {
