+++ src/view/popup/popup.tsx
@@ -4,7 +4,7 @@ import { CountryDropdown } from "../components/country-dropdown/CountryDropdown"
 import { usePopup } from "./usePopup";
 import { SelectedCountries } from "../components/selected-countries/SelectedCountries";
 import { CalculationHistory } from "../components/calculation-history/CalculationHistory";
-import { formatBytesString } from "../../utils/helpers/formatBytesString";
+import { formatBytesString } from "./utils/formatBytesString";
 
 export const Popup = () => {
     const {
