+++ src/view/popup/utils/formatBytesString.ts
